# openapi.yaml
openapi: 3.0.3
info:
  title: ClawdBridge API
  description: Mobile bridge for Claude Code — control your desktop AI agent from anywhere
  version: 1.0.0
  contact:
    name: ClawdBridge Team
servers:
  - url: http://localhost:4338/api/v1
    description: Local development
  - url: https://api.clawdbridge.io/api/v1
    description: Production

paths:
  /auth/oauth:
    post:
      summary: OAuth login
      description: Authenticate with GitHub OAuth and receive JWT token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, code]
              properties:
                provider:
                  type: string
                  enum: [github]
                code:
                  type: string
                  description: GitHub OAuth authorization code
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/refresh:
    post:
      summary: Refresh token
      description: Obtain a new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  /tasks:
    get:
      summary: List tasks
      description: Get all tasks for authenticated user with optional filtering
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, paused, completed, error]
        - name: repo
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

    post:
      summary: Create task
      description: Create a new coding task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, repo]
              properties:
                title:
                  type: string
                repo:
                  type: string
                category:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /tasks/{taskId}:
    get:
      summary: Get task
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'

  /tasks/{taskId}/sessions:
    post:
      summary: Create session
      description: Create a new chat session for a task
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    $ref: '#/components/schemas/Session'

  /sessions/{sessionId}/messages:
    get:
      summary: Get messages
      description: Get messages for a session with pagination
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: integer
            description: Timestamp filter
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'

  /files:
    post:
      summary: Upload file
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  fileId:
                    type: string

  /files/{fileId}:
    get:
      summary: Download file
      security:
        - bearerAuth: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /metrics:
    get:
      summary: Prometheus metrics
      responses:
        '200':
          description: Metrics in Prometheus format
          content:
            text/plain:
              schema:
                type: string

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        login:
          type: string
        avatar_url:
          type: string

    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        repo:
          type: string
        status:
          type: string
          enum: [pending, in_progress, paused, completed, error]
        tags:
          type: array
          items:
            type: string
        category:
          type: string
        createdAt:
          type: integer
        updatedAt:
          type: integer

    Session:
      type: object
      properties:
        id:
          type: string
        taskId:
          type: string
        title:
          type: string
        status:
          type: string
        workDir:
          type: string
        createdAt:
          type: integer
        lastMessageAt:
          type: integer

    Message:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        type:
          type: string
          enum: [user, assistant, tool_call, tool_result, error, system]
        content:
          type: string
        seq:
          type: integer
        timestamp:
          type: integer
        status:
          type: string
        metadata:
          type: object

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
              reqId:
                type: string

    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string

    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string